Sales_details
This is a sanitised version of an algorithm that sources the Sales person from Client data. It provides information of the Sales person. Developed in Teradata environment.

The final table output show:

Datasource - Where the information for the sales person was sourced from
Name of sales
DOB of sales
Balance of sales